+++ sela/data/dataset.py
@@ -9,8 +9,8 @@ import pandas as pd
 import yaml
 from sklearn.model_selection import train_test_split
 
-from expo.insights.solution_designer import SolutionDesigner
-from expo.utils import DATA_CONFIG
+from sela.insights.solution_designer import SolutionDesigner
+from sela.utils import DATA_CONFIG
 
 BASE_USER_REQUIREMENT = """
 This is a {datasetname} dataset. Your goal is to predict the target column `{target_col}`.
