+++ sela/evaluation/visualize_mcts.py
@@ -3,7 +3,7 @@ import textwrap
 import matplotlib.pyplot as plt
 import networkx as nx
 
-from expo.MCTS import Node
+from sela.MCTS import Node
 
 NODE_TEMPLATE = """\
 [Node {id}]
