+++ sela/experimenter/autogluon.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from expo.experimenter.custom import CustomExperimenter
+from sela.experimenter.custom import CustomExperimenter
 import os
 import pandas as pd
