+++ sela/scripts/run_reg.sh
